♻️ Refactor: 공통 로직 및 UI 분리
